+++ src/main.tsx
@@ -5,6 +5,18 @@
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import '@fontsource/inter/latin-400.css';
+import '@fontsource/inter/latin-400-italic.css';
+import '@fontsource/inter/latin-500.css';
+import '@fontsource/inter/latin-500-italic.css';
+import '@fontsource/inter/latin-600.css';
+import '@fontsource/inter/latin-600-italic.css';
+import '@fontsource/inter/latin-700.css';
+import '@fontsource/inter/latin-700-italic.css';
+import '@fontsource/inter/latin-800.css';
+import '@fontsource/inter/latin-800-italic.css';
+import '@fontsource/inter/latin-900.css';
+import '@fontsource/inter/latin-900-italic.css';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';